일부 기능 수정
