+++ src/pages/board/Board.module.css
@@ -6,19 +6,12 @@
 /* 공지사항 */
 .notice_all {display: flex; width: 1180px; padding-top: 20px; position: relative; margin: 18px auto;}
 .noticeTitle {font-size: 22px; padding-left: 20px;}
+.notice_head {font-weight: bold; font-size: 16px; color: #000000;}
+.notice_toggle {font-size: 18px; background-color: #ffffff; border: none; height: 42px; color: #999999;}
+.notice_content {margin-left: 20px; font-size: 14px; color: #000000;}
 
 /* 공지 토글*/
 .toggle {width: 1180px; margin: 0 auto;}
-.notice_toggle {font-size: 18px; background-color: #ffffff; border: none; height: 42px; color: #999999;}
-.notice_head {font-size: 16px; color: #999999;}
-.notice_content {font-size: 14px; color: #999999;}
-
-/* 딱히 쓸모없는거 확인하고 지울거
-.noticeBox 
-    {width: 1180px; height: 56px; color: #999999;
-    border-top: solid 1px#D4D4D4; border-bottom:
-    solid 1px #D4D4D4;
-    }*/
 
 /* 게시판 */
 .board_title {display: flex; align-items: baseline; justify-content: space-between; width: 1180px; position: relative; margin: 0px auto; padding-top: 32px; padding-left: 20px; font-size: 22px;}
@@ -70,9 +63,13 @@
 /* 타이틀*/
 .boxTitle
 {
-    padding-left: 14px;
+    padding-left: 20px;
     font-size: 14px;
     color: #999999;
+    text-overflow: ellipsis;
+    overflow: hidden;
+    width: 1100px;
+    white-space: nowrap;
 }
 /* 닉네임 */
 .nick
